requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.2.0
pydicom>=3.0.0
SimpleITK>=2.3.0
itk>=5.4.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0
pillow>=10.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
python-docx>=1.0.0
openpyxl>=3.1.0

# Logging & Experiment Tracking
tensorboard>=2.14.0

# Optional: Explainability (uncomment if needed)
# shap>=0.42.0
